revert: put pipeline guessing back to CLI
