+++ bin/barnard59.js
@@ -4,7 +4,15 @@ const fs = require('fs')
 const p = require('..')
 const path = require('path')
 const program = require('commander')
+const cf = require('clownface')
+const rdf = require('rdf-ext')
+const namespace = require('@rdfjs/namespace')
 const runner = require('../lib/runner')
+
+const ns = {
+  p: namespace('https://pipeline.described.at/'),
+  rdf: namespace('http://www.w3.org/1999/02/22-rdf-syntax-ns#')
+}
 
 function createOutputStream (output) {
   if (output === '-') {
@@ -24,6 +32,32 @@ function parseVariables (str, all) {
     }, all)
 }
 
+function guessPipeline (dataset) {
+  const graph = cf(dataset)
+
+  const pipelines = graph.has(ns.rdf('type'), [ ns.p('Pipeline'), ns.p('ObjectPipeline') ])
+
+  if (pipelines.values.length === 0) {
+    throw new Error('no pipeline found in the dataset')
+  }
+
+  const rootPipelines = pipelines.values.reduce((arr, id) => {
+    const node = dataset.match(null, null, rdf.namedNode(id), null)
+
+    if (node.length === 0) {
+      arr.push(id)
+    }
+
+    return arr
+  }, [])
+
+  if (rootPipelines.length > 1) {
+    throw new Error('multiple root pipeline found. please specify the one to run using --pipeline option')
+  }
+
+  return rootPipelines[0]
+}
+
 program
   .command('run <filename>')
   .option('--format <mediaType>', 'media type of the pipeline description', 'application/ld+json')
@@ -32,14 +66,23 @@ program
   .option('--variable <name=value>', 'variable key value pairs separated by comma', parseVariables, new Map())
   .option('-v, --verbose', 'enable diagnostic console output')
   .action((filename, options = {}) => {
-    const { format, output } = options
+    let { format, output, pipeline } = options
     p.fileToDataset(format, filename)
-      .then(runner.create({
-        ...options,
-        outputStream: createOutputStream(output),
-        log: process.stdout,
-        basePath: path.resolve(path.dirname(filename))
-      }))
+      .then(dataset => {
+        if (!pipeline) {
+          pipeline = guessPipeline(dataset)
+        }
+
+        const run = runner.create({
+          ...options,
+          pipeline,
+          outputStream: createOutputStream(output),
+          log: process.stdout,
+          basePath: path.resolve(path.dirname(filename))
+        })
+
+        return run(dataset)
+      })
       .catch(err => {
         console.error(err)
         process.exit(1)
